Move resources.utils to pages app
